Added Authentication on AddBill Method
